requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
sqlalchemy
pgvector
python-multipart
openai
httpx
tiktoken
python-docx
PyPDF2
firebase-admin
google-cloud-storage

langgraph
langgraph-checkpoint-postgres
langchain-openai
psycopg[binary,pool]

pytest
